update: dropdown-session details
